Add some caching of known information. Fix a recursion problem, and the broken test fixtures.
No need to generate an annotation index for every entity type, so make
`IndexFactory` keep track of what it has indexed and return previously
indexed types when requested.

The Marshaller is now a little more structured and organized with the
introduction of a Subresource class which manages reflective invocation
of a subresources method on an entity.

The test fixtures package names were all wrong.