Fix enumeration generation for enums starting with a number.
* Motivation:

For DMR configuration values that become Java enums, those that
start with a number are no legal enum values.  For instance, this
is invalid:

enum Foo {
  1_FOO,
  2_FOO
}

This would occur when a DMR value was "1-foo" or "2-foo".

* Modifications:

We now provide exceptions for enum values that start with digits.
By default, if the leading character of an enum value is a digit,
we prepend a literal underscore, resulting in

enum Foo {
  _1_FOO,
  _2_FOO
  _99_FOO
}

Additionally, we perform another pass for those enums that have a
single digit at the head, and replace it with the english equivalent,
which would result in

enum Foo {
  ONE_FOO,
  TWO_FOO,
  _99_FOO
}

The idea being that for single digits, we can avoid the ugliness of
the prepended underscore, as the only case we've seen thus far is
2_WAY and 3_WAY which become TWO_WAY and THREE_WAY.

* Result:

Teiid's config-api now generates compilable code.