[SWARM-1987] properly support composite types in Config API
Motivation
----------
To be able to fully configure Elytron, which relies on nesting
`Map`s into `List`s which are themselves nested inside `Map`s,
the `ListTypeAdapter` and `MapTypeAdapter` must be able to
convert composite types too.

Modifications
-------------
This is easiest to achieve when the adapters can delegate
to each other. To enable that, they no longer take a `ModelNode`
and a name which should be looked up in that `ModelNode`; instead,
they take the target `ModelNode` directly. This makes it natural
for each of the type adapters to delegate to the other ones.

So now:
- `SimpleTypeAdapter` handles conversion of non-composite types
- `ListTypeAdapter` handles conversion of lists and delegates to
  `SimpleTypeAdapter` for basic types and to `MapTypeAdapter`
   for maps
- `MapTypeAdapter` handles conversion of maps and delegates to
  `SimpleTypeAdapter` for basic types and to `ListTypeAdapter`
  for lists

This is naturally recursive, where `SimpleTypeAdapter` is the base
case, and it's guaranteed to complete by construction. The only case
when this procedure wouldn't complete is when the converted structure
would itself be infinite.

Result
------

First and foremost, able to fully configure Elytron.
As a nice side effect, it's now possible to put non-`String` values
into `Map`s and get correct DMR output.